frontend/requirements.txt
streamlit
requests
streamlit-oauth
python-dotenv
